move LIGHT_TYPES dict into module that can be imported by hydra and non-hydra test files, disable the Decal and Grid tests as they need their own test scripts (this one is too cluttered)
